CLI for stage 1 and relationships for stage 2

+++ src/main/java/org/example/repositories/MemberRepository.java
@@ -51,8 +51,9 @@ public class MemberRepository {
 
         session.beginTransaction();
 
-        session.delete(member);
-
+        if(member != null){
+            session.delete(member);
+        }
         session.getTransaction().commit();
 
         session.close();
